Update installation config based on learnings from new machine set up
Automate rbenv installation and use it to install a modern ruby version.
Also install rbenv plugins rbenv-ctags and rbenv-default-gems that I've
depended on in the past.

Although rbenv-default-gems was installed, it didn't behave as expected
as my default gems did not get installed when installing a new ruby
version with rbenv, so something isn't quite right, but this is an
improvement over what I had before.